Change carousel and fix side indexes
The slider was built by hand and presented a few problems for both functioning and responsivity, making it hard to mantain under the short amount of time for the feature, now the app uses react-slick, a responsive carousel slider which makes the job easier and faster. On top of that, this commit also fixes some css problems related to the side indexes.